backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.58.1
pydantic==2.10.4
httpx==0.28.1
python-multipart==0.0.20

